# String.Join optimization for single item lists