Returns existing string instance for `String.Join` when there is only one item in the list. Optimizations for when `IEnumerable<string>` is a `IList<string>`.
I've tried to match the coding style in this file but it is a bit variable. CoreFX `System.Runtime` tests have been run on these changes.

It is common to have lists that may have multiple items but typically have only one. Running a String.Join on them will needlessly create a new string.
One example is http header parsing.

https://github.com/aspnet/HttpAbstractions/blob/dev/src/Microsoft.AspNet.Http.Extensions/ParsingHelpers.cs#L537

Some headers support multi values but will usually only have one. In the AspNet implementation all headers are in a `IEnumerable` to simplify the interfaces.

This PR is mainly about reducing allocations but there are some immediate performance benefits to go along with it. Perf tests done with 1M iterations.

### String[] Join
- Early exit when there is only one item. Avoids new string allocation. This is basically the reverse of https://github.com/dotnet/coreclr/commit/11cddd0eff920152206345acfa827df9fa855ecc#diff-d2a641ec9b77a5ea1cc985bdec3e74ef
- Store `value[stringToJoinIndex]` in a local variable to avoid double array lookup.

| Test value | Before | After | Faster |
| ---------- | -----: | ----: | -----: |
| string[1] | 46 | 13 | 254% |
| string[5] | 232 | 225 | 3% |
| string[100] | 4683 | 4564 | 3% |

### `IEnumerable<String>` Join 
- Removed null separator normalisation to `String.Empty`. `StringBuilder.Append` handles nulls and is the first check in that function. So passing null is actually more desirable.

### `IEnumerable<String>` Join when values is an `IList<String>`
- Use `IList` interface for Count and index property access. Same optimisation as many of the Linq methods. Early exit for when Count is 0 and 1, avoids new string allocation.
- Use for loop for iteration to avoid enumerator creation

| Test value | Before | After | Faster |
| ---------- | -----: | ----: | -----: |
| List<string>(1) | 89 | 14 | 538% |
| List<string>(5) | 212 | 145 | 46% |
| List<string>(100) | 3570 | 2565 | 39% |

### `IEnumerable<String>` Join when values is not an `IList<String>` (used `Queue<String>` for testing)
- Removed null checks on `en.Current`. As above, `Append` handles nulls and this removes the double property access and branch for what should be a rarer case.

| Test value | Before | After | Faster |
| ---------- | -----: | ----: | -----: |
| Queue<string>(1) | 95 | 45 | 113% |
| Queue<string>(5) | 240 | 182 | 31% |
| Queue<string>(100) | 4021 | 3236 | 24% |

Edit: `IList<string>` path is no longer included in this PR