@Eyas 2015-08-30
Since you are only depending on the read-only properties of `IList`, is it better to make this `IReadOnlyList`-specific instead of `IList`-specific?

This means we will see these improvements for `List<T>` (which implements `IReadOnlyList<T>`), as well as `ImmutableList<T>` and `ImmutableArray<T>`, etc.

This means there would be no slow-down for *some* custom mutable list-like implementations that only inherit from `IList`. Still, might be worthwhile.