@sharwell 2015-08-30
@Eyas `ImmutableList<T>` and `ImmutableArray<T>` both implement `IList<T>`.