@Eyas 2015-08-30
Ah yes, bad example. But my suggestion stands. Why not use the minimal interface that provides only the functions needed to get the perf boost?

The alternative is to force an implementer to always implement IList to be treated as a first class citizen, even of the mutating methods are throw-only