@bbowyersmyth 2015-08-31
Interesting suggestion @Eyas I used IList mainly for consistency with other areas such as Linq and I'm not sure how popular IReadOnlyList has become yet.

Question for the maintainers, any preferences on this going forward?