@ayende 2015-08-31
Aren't there implications about interface calls vs. direct call here, WRT
performance?

*Hibernating Rhinos Ltd  *

Oren Eini* l CEO l *Mobile: + 972-52-548-6969

Office: +972-4-622-7811 *l *Fax: +972-153-4-622-7811



On Mon, Aug 31, 2015 at 5:53 AM, Bruce Bowyer-Smyth <
notifications@github.com> wrote:

> Interesting suggestion @Eyas <https://github.com/Eyas> I used IList
> mainly for consistency with other areas such as Linq and I'm not sure how
> popular IReadOnlyList has become yet.
>
> Question for the maintainers, any preferences on this going forward?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/dotnet/coreclr/pull/1460#issuecomment-136239306>.
>