@JonHanna 2015-08-31
There's a performance difference in the test of `as ISomeInterface<T>` when `ISomeInterface` is variant and when it isn't, presumably due to the fact that a given object could be implementing such an interface or also the same interface but for a type that works with the variance.
`IReadOnlyList<T>` is covariant, so it will suffer here. Whether it suffers enough to outweigh the possible benefits would need measurements.