@bbowyersmyth 2015-08-31
@ayende forcing List into either path the timings are:

| Test Value | IList | IEnumerable |
| ------ | -----: | ------: |
| List(1) | 14 | 14 |
| List(5) | 145 | 195 |
| List(100) | 2565 | 3327 |

If `list.Count` was cached into a local variable to avoid multiple interface calls to that:

| Test Value | IList |
| ------ | -----: |
| List(1) | 14 |
| List(5) | 127 |
| List(100) | 2267 |

So that might be worth doing which would leave one `Count` and multiple indexer accesses.
Was there a test you had in mind?