@ayende 2015-08-31
No, I meant List<T> vs. IList<T> vs. IReadOnlyList<T>

*Hibernating Rhinos Ltd  *

Oren Eini* l CEO l *Mobile: + 972-52-548-6969

Office: +972-4-622-7811 *l *Fax: +972-153-4-622-7811



On Mon, Aug 31, 2015 at 3:20 PM, Bruce Bowyer-Smyth <
notifications@github.com> wrote:

> @ayende <https://github.com/ayende> forcing List into either path the
> timings are:
> Test Value IList IEnumerable List(1) 14 14 List(5) 145 195 List(100) 2565
> 3327
>
> If list.Count was cached into a local variable to avoid multiple
> interface calls to that:
> Test Value IList List(1) 14 List(5) 127 List(100) 2267
>
> So that might be worth doing which would leave one Count and multiple
> indexer accesses.
> Was there a test you had in mind?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/dotnet/coreclr/pull/1460#issuecomment-136356604>.
>