@JonHanna 2015-08-31
I'm curious as to what results you get if you have only the improvement for the single-element enumerable case, and remove the `IList<string>` path entirely. In particular the effect when the `IList<string>` in question is `string[]`.