@bbowyersmyth 2015-08-31
With the `list.Count` caching change.

| Test Value | List&lt;T&gt; | IReadOnlyList&lt;T&gt; | No IList path at all |
| ------ | -----: | -----: | -----: |
| List(1) | 7 | 131 | 37 |
| List(5) | 112 | 248 | 196 |
| List(100) | 1987 | 2342 | 3274 |

There is a specific Join overload for string[] but if the value arrived as an IList&lt;T&gt;

| Test Value | Before (existing) | IList path | No IList path at all |
| ------ | -----: | -----: | -----: |
| (IList&lt;T&gt;) string[1] | 83 | 43 | 28 |
| (IList&lt;T&gt;) string[5] | 182 | 165 | 164 |
| (IList&lt;T&gt;) string[100] | 3336 | 2455 | 2753 |