@bbowyersmyth 2015-09-04
Committed the count local variable cache. The IList&lt;string&gt; try cast still seems like the best general improvement so we go with that or play it safe and just stick with the IEnumerable path.

Gist for perf testing:
https://gist.github.com/bbowyersmyth/c1c55444fb102281e3c9