@bbowyersmyth 2015-09-07
I'm thinking that the code should just stick with IEnumerable path for now (which is still a win). There is no guarantee that IList indexers have to return values in the same order as their enumerator. While it is probably fine for Linq to work on that assumption, `System.String` should be more correct.
Unless there are any other thoughts I'll resubmit this commit later this afternoon.