@bbowyersmyth 2015-09-07
Replaced the commit to remove the IList try cast.