@bbowyersmyth 2015-09-10
Addressed the feedback with variable name, nesting, and null handling comment