@AlexGhiondea 2015-09-11
:+1: 

Thank you @bbowyersmyth for the change! If you don't mind, could you squash the commit that adressed the PR feedback? 

I prefer having separate commits while the PR is open as that makes it easier to see what has changes after each feedback iteration but I don't think they add value once we are done with the feedback :).