@AlexGhiondea 2015-09-11
Thanks @bbowyersmyth !

Waiting for a second :+1: in order to merge it in!