@stephentoub 2015-09-16
LGTM.  Do we have tests in corefx for these special cases?