@bbowyersmyth 2015-09-16
@stephentoub Yes they were added via https://github.com/dotnet/corefx/pull/2839